refactor: clean up getWorkExperience Action
